Split views into three (admin/update/search) for navigability.
https://trello.com/c/5oIKvS1s/100-search-service-for-briefs-plus-indexing-script
