+++ app/main/__init__.py
@@ -6,4 +6,4 @@ main = Blueprint('main', __name__)
 main.before_request(requires_authentication)
 
 from .errors import *
-from app.main.views import search, meta
+from app.main.views import search, meta, admin, update
