+++ tests/app/helpers.py
@@ -109,3 +109,7 @@ def assert_response_status(response, expected_status):
     assert response.status_code == expected_status, "Expected {} response; got {}. {}".format(
         expected_status, response.status_code, response.get_data(as_text=True)
     )
+
+
+def get_json_from_response(response):
+    return json.loads(response.get_data())
