+++ tests/app/views/test_search.py
@@ -1,262 +1,17 @@
 import mock
 import pytest
 from flask import json
-from werkzeug import MultiDict
 from nose.tools import assert_equal, assert_in, assert_true
-from urllib3.exceptions import NewConnectionError
+from werkzeug import MultiDict
 
 from app import elasticsearch_client as es
-from app.main.services.search_service import core_search_and_aggregate
 from app.main.services import search_service
-
-from ..helpers import (BaseApplicationTest, BaseApplicationTestWithIndex, assert_response_status,
+from app.main.services.search_service import core_search_and_aggregate
+from tests.app.helpers import get_json_from_response
+from ..helpers import (BaseApplicationTestWithIndex, assert_response_status,
                        make_search_api_url, make_standard_service)
 
-
 pytestmark = pytest.mark.usefixtures("services_mapping")
-
-
-class TestSearchIndexes(BaseApplicationTest):
-    def test_should_be_able_create_and_delete_index(self):
-        response = self.create_index()
-        assert_response_status(response, 200)
-        assert_equal(get_json_from_response(response)["message"],
-                     "acknowledged")
-
-        response = self.client.get('/index-to-create')
-        assert_response_status(response, 200)
-
-        response = self.client.delete('/index-to-create')
-        assert_response_status(response, 200)
-        assert_equal(get_json_from_response(response)["message"],
-                     "acknowledged")
-
-        response = self.client.get('/index-to-create')
-        assert_response_status(response, 404)
-
-    def test_should_be_able_to_create_aliases(self):
-        self.create_index()
-        response = self.client.put('/index-alias', data=json.dumps({
-            "type": "alias",
-            "target": "index-to-create"
-        }), content_type="application/json")
-
-        assert_response_status(response, 200)
-        assert_equal(get_json_from_response(response)["message"], "acknowledged")
-
-    def test_should_not_be_able_to_delete_aliases(self):
-        self.create_index()
-        self.client.put('/index-alias', data=json.dumps({
-            "type": "alias",
-            "target": "index-to-create"
-        }), content_type="application/json")
-
-        response = self.client.delete('/index-alias')
-
-        assert_response_status(response, 400)
-        assert_equal(get_json_from_response(response)["error"], "Cannot delete alias 'index-alias'")
-
-    def test_should_not_be_able_to_delete_index_with_alias(self):
-        self.create_index()
-        self.client.put('/index-alias', data=json.dumps({
-            "type": "alias",
-            "target": "index-to-create"
-        }), content_type="application/json")
-
-        response = self.client.delete('/index-to-create')
-
-        assert_response_status(response, 400)
-        assert_equal(
-            get_json_from_response(response)["error"],
-            "Index 'index-to-create' is aliased as 'index-alias' and cannot be deleted"
-        )
-
-    def test_cant_create_alias_for_missing_index(self):
-        response = self.client.put('/index-alias', data=json.dumps({
-            "type": "alias",
-            "target": "index-to-create"
-        }), content_type="application/json")
-
-        assert_response_status(response, 404)
-        assert get_json_from_response(response)["error"].startswith(
-            'index_not_found_exception: no such index')
-
-    def test_cant_replace_index_with_alias(self):
-        self.create_index()
-        response = self.client.put('/index-to-create', data=json.dumps({
-            "type": "alias",
-            "target": "index-to-create"
-        }), content_type="application/json")
-
-        assert_response_status(response, 400)
-        assert_equal(get_json_from_response(response)["error"],
-                     'invalid_alias_name_exception: Invalid alias name [index-to-create], an index exists with the '
-                     'same name as the alias (index-to-create)')
-
-    def test_can_update_alias(self):
-        self.create_index()
-        self.create_index('index-to-create-2')
-        self.client.put('/index-alias', data=json.dumps({
-            "type": "alias",
-            "target": "index-to-create"
-        }), content_type="application/json")
-
-        response = self.client.put('/index-alias', data=json.dumps({
-            "type": "alias",
-            "target": "index-to-create-2"
-        }), content_type="application/json")
-
-        assert_response_status(response, 200)
-        status = get_json_from_response(self.client.get('/_all'))["status"]
-        assert_equal(status['index-to-create']['aliases'], [])
-        assert_equal(status['index-to-create-2']['aliases'], ['index-alias'])
-
-    def test_creating_existing_index_updates_mapping(self):
-        self.create_index()
-
-        with self.app.app_context():
-            with mock.patch(
-                'app.main.services.search_service.es.indices.put_mapping'
-            ) as es_mock:
-                response = self.create_index()
-
-        assert_response_status(response, 200)
-        assert_equal("acknowledged", get_json_from_response(response)["message"])
-        es_mock.assert_called_with(
-            index='index-to-create',
-            doc_type='services',
-            body=mock.ANY
-        )
-
-    def test_should_not_be_able_delete_index_twice(self):
-        self.create_index()
-        self.client.delete('/index-to-create')
-        response = self.client.delete('/index-to-create')
-        assert_response_status(response, 404)
-        assert_equal(get_json_from_response(response)["error"],
-                     'index_not_found_exception: no such index (index-to-create)')
-
-    def test_should_return_404_if_no_index(self):
-        response = self.client.get('/index-does-not-exist')
-        assert_response_status(response, 404)
-        assert_equal(get_json_from_response(response)["error"],
-                     "index_not_found_exception: no such index (index-does-not-exist)")
-
-    def test_bad_mapping_name_gives_400(self):
-        response = self.client.put('/index-to-create', data=json.dumps({
-            "type": "index",
-            "mapping": "some-bad-mapping"
-        }), content_type="application/json")
-
-        assert_response_status(response, 400)
-        assert get_json_from_response(response)["error"] == "Mapping definition named 'some-bad-mapping' not found."
-
-
-class TestIndexingDocuments(BaseApplicationTestWithIndex):
-
-    EXAMPLE_CONNECTION_ERROR = (
-        '<urllib3.connection.HTTPConnection object at 0x107626588>: '
-        'Failed to establish a new connection: [Errno 61] Connection refused'
-    )
-
-    def setup(self):
-        super(TestIndexingDocuments, self).setup()
-        self.create_index()
-
-    def test_should_index_a_document(self, default_service):
-        service = default_service
-
-        response = self.client.put(
-            make_search_api_url(service),
-            data=json.dumps(service),
-            content_type='application/json')
-
-        assert_response_status(response, 200)
-
-        with self.app.app_context():
-            search_service.refresh('index-to-create')
-        response = self.client.get('/index-to-create')
-        assert_response_status(response, 200)
-        assert_equal(
-            get_json_from_response(response)["status"]["num_docs"],
-            1)
-
-    @mock.patch('app.main.views.search.index', return_value=(NewConnectionError('', EXAMPLE_CONNECTION_ERROR), 'N/A'))
-    def test_index_document_handles_connection_error(self, ind, default_service):
-        service = default_service
-
-        response = self.client.put(
-            make_search_api_url(service),
-            data=json.dumps(service),
-            content_type='application/json'
-        )
-        assert_response_status(response, 500)
-
-    @mock.patch(
-        'app.main.views.search.index',
-        return_value=(NewConnectionError('', EXAMPLE_CONNECTION_ERROR), 'something_other_than_N/A')
-    )
-    def test_index_document_does_not_pass_on_non_NA_status_code(self, ind, default_service):
-        service = default_service
-        with pytest.raises(TypeError):
-            self.client.put(
-                make_search_api_url(service),
-                data=json.dumps(service),
-                content_type='application/json'
-            )
-
-    def test_should_index_a_document_with_missing_fields(self, default_service):
-        service = default_service
-        del service.get('document', service.get('service'))["serviceName"]
-
-        response = self.client.put(
-            make_search_api_url(service),
-            data=json.dumps(service),
-            content_type='application/json')
-
-        assert_response_status(response, 200)
-
-    def test_should_index_a_document_with_extra_fields(self, default_service):
-        service = default_service
-        service.get('document', service.get('service'))["randomField"] = "some random"
-
-        response = self.client.put(
-            make_search_api_url(service),
-            data=json.dumps(service),
-            content_type='application/json')
-
-        assert_response_status(response, 200)
-
-    def test_should_index_a_document_with_incorrect_types(self, default_service):
-        service = default_service
-        service.get('document', service.get('service'))["serviceName"] = 123
-
-        response = self.client.put(
-            make_search_api_url(service),
-            data=json.dumps(service),
-            content_type='application/json')
-
-        assert_response_status(response, 200)
-
-    def test_should_index_a_document_with_no_service_types(self, default_service):
-        service = default_service
-        service.get('document', service.get('service'))["serviceName"] = 123
-        del service.get('document', service.get('service'))["serviceTypes"]
-
-        response = self.client.put(
-            make_search_api_url(service),
-            data=json.dumps(service),
-            content_type='application/json')
-
-        assert_response_status(response, 200)
-
-    def test_should_raise_400_on_bad_doc_type(self, default_service):
-        response = self.client.put(
-            make_search_api_url(default_service, type_name='some-bad-type'),
-            data=json.dumps(default_service),
-            content_type='application/json')
-
-        assert_response_status(response, 400)
 
 
 @pytest.mark.usefixtures("services_mapping_definition")
@@ -550,39 +305,6 @@ class TestFetchById(BaseApplicationTestWithIndex):
                 service.get('document', service.get('service'))[key])  # TODO remove shim
 
 
-class TestDeleteById(BaseApplicationTestWithIndex):
-    def test_should_delete_service_by_id(self, default_service):
-        service = default_service
-        self.client.put(
-            make_search_api_url(service),
-            data=json.dumps(service),
-            content_type='application/json'
-        )
-
-        with self.app.app_context():
-            search_service.refresh('index-to-create')
-        response = self.client.delete(make_search_api_url(service),)
-
-        data = get_json_from_response(response)
-        assert_response_status(response, 200)
-        assert_equal(data['message']['found'], True)
-
-        response = self.client.get(make_search_api_url(service),)
-        data = get_json_from_response(response)
-        assert_response_status(response, 404)
-        assert_equal(data['error']['found'], False)
-
-    def test_should_return_404_if_no_service(self):
-        self.create_index()
-
-        response = self.client.delete(
-            '/index-to-create/delete/100')
-
-        data = get_json_from_response(response)
-        assert_response_status(response, 404)
-        assert_equal(data['error']['found'], False)
-
-
 class TestSearchType(BaseApplicationTestWithIndex):
     def test_core_search_and_aggregate_does_dfs_query_for_searches(self):
         with self.app.app_context(), mock.patch.object(es, 'search') as es_search_mock:
@@ -626,7 +348,3 @@ def create_services(number_of_services):
         services.append(service)
 
     return services
-
-
-def get_json_from_response(response):
-    return json.loads(response.get_data())
